requirements.txt
aiogram==3.11.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
alembic==1.13.1
aiosmtplib==2.0.2
httpx==0.27.0
chromadb==0.5.3
sentence-transformers==2.5.1

# Явно указываем CPU-версию PyTorch
torch --index-url https://download.pytorch.org/whl/cpu
